requirements.txt
numpy
matplotlib
scikit-image
ipython
imageio
pillow
scipy
tqdm
